@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v3.19.6
+// 	protoc-gen-go v1.35.1
+// 	protoc        v5.28.3
 // source: geyser.proto
 
 package proto
@@ -88,11 +88,9 @@ type SubscribeRequest struct {
 
 func (x *SubscribeRequest) Reset() {
 	*x = SubscribeRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequest) String() string {
@@ -103,7 +101,7 @@ func (*SubscribeRequest) ProtoMessage() {}
 
 func (x *SubscribeRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -200,11 +198,9 @@ type SubscribeRequestFilterAccounts struct {
 
 func (x *SubscribeRequestFilterAccounts) Reset() {
 	*x = SubscribeRequestFilterAccounts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestFilterAccounts) String() string {
@@ -215,7 +211,7 @@ func (*SubscribeRequestFilterAccounts) ProtoMessage() {}
 
 func (x *SubscribeRequestFilterAccounts) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -266,11 +262,9 @@ type SubscribeRequestFilterAccountsFilter struct {
 
 func (x *SubscribeRequestFilterAccountsFilter) Reset() {
 	*x = SubscribeRequestFilterAccountsFilter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestFilterAccountsFilter) String() string {
@@ -281,7 +275,7 @@ func (*SubscribeRequestFilterAccountsFilter) ProtoMessage() {}
 
 func (x *SubscribeRequestFilterAccountsFilter) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -364,11 +358,9 @@ type SubscribeRequestFilterAccountsFilterMemcmp struct {
 
 func (x *SubscribeRequestFilterAccountsFilterMemcmp) Reset() {
 	*x = SubscribeRequestFilterAccountsFilterMemcmp{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestFilterAccountsFilterMemcmp) String() string {
@@ -379,7 +371,7 @@ func (*SubscribeRequestFilterAccountsFilterMemcmp) ProtoMessage() {}
 
 func (x *SubscribeRequestFilterAccountsFilterMemcmp) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -464,11 +456,9 @@ type SubscribeRequestFilterSlots struct {
 
 func (x *SubscribeRequestFilterSlots) Reset() {
 	*x = SubscribeRequestFilterSlots{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestFilterSlots) String() string {
@@ -479,7 +469,7 @@ func (*SubscribeRequestFilterSlots) ProtoMessage() {}
 
 func (x *SubscribeRequestFilterSlots) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -516,11 +506,9 @@ type SubscribeRequestFilterTransactions struct {
 
 func (x *SubscribeRequestFilterTransactions) Reset() {
 	*x = SubscribeRequestFilterTransactions{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestFilterTransactions) String() string {
@@ -531,7 +519,7 @@ func (*SubscribeRequestFilterTransactions) ProtoMessage() {}
 
 func (x *SubscribeRequestFilterTransactions) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -601,11 +589,9 @@ type SubscribeRequestFilterBlocks struct {
 
 func (x *SubscribeRequestFilterBlocks) Reset() {
 	*x = SubscribeRequestFilterBlocks{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestFilterBlocks) String() string {
@@ -616,7 +602,7 @@ func (*SubscribeRequestFilterBlocks) ProtoMessage() {}
 
 func (x *SubscribeRequestFilterBlocks) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -667,11 +653,9 @@ type SubscribeRequestFilterBlocksMeta struct {
 
 func (x *SubscribeRequestFilterBlocksMeta) Reset() {
 	*x = SubscribeRequestFilterBlocksMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestFilterBlocksMeta) String() string {
@@ -682,7 +666,7 @@ func (*SubscribeRequestFilterBlocksMeta) ProtoMessage() {}
 
 func (x *SubscribeRequestFilterBlocksMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -705,11 +689,9 @@ type SubscribeRequestFilterEntry struct {
 
 func (x *SubscribeRequestFilterEntry) Reset() {
 	*x = SubscribeRequestFilterEntry{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestFilterEntry) String() string {
@@ -720,7 +702,7 @@ func (*SubscribeRequestFilterEntry) ProtoMessage() {}
 
 func (x *SubscribeRequestFilterEntry) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -746,11 +728,9 @@ type SubscribeRequestAccountsDataSlice struct {
 
 func (x *SubscribeRequestAccountsDataSlice) Reset() {
 	*x = SubscribeRequestAccountsDataSlice{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestAccountsDataSlice) String() string {
@@ -761,7 +741,7 @@ func (*SubscribeRequestAccountsDataSlice) ProtoMessage() {}
 
 func (x *SubscribeRequestAccountsDataSlice) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -800,11 +780,9 @@ type SubscribeRequestPing struct {
 
 func (x *SubscribeRequestPing) Reset() {
 	*x = SubscribeRequestPing{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[10]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeRequestPing) String() string {
@@ -815,7 +793,7 @@ func (*SubscribeRequestPing) ProtoMessage() {}
 
 func (x *SubscribeRequestPing) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[10]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -859,11 +837,9 @@ type SubscribeUpdate struct {
 
 func (x *SubscribeUpdate) Reset() {
 	*x = SubscribeUpdate{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[11]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[11]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdate) String() string {
@@ -874,7 +850,7 @@ func (*SubscribeUpdate) ProtoMessage() {}
 
 func (x *SubscribeUpdate) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[11]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1036,11 +1012,9 @@ type SubscribeUpdateAccount struct {
 
 func (x *SubscribeUpdateAccount) Reset() {
 	*x = SubscribeUpdateAccount{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[12]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[12]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateAccount) String() string {
@@ -1051,7 +1025,7 @@ func (*SubscribeUpdateAccount) ProtoMessage() {}
 
 func (x *SubscribeUpdateAccount) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[12]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1104,11 +1078,9 @@ type SubscribeUpdateAccountInfo struct {
 
 func (x *SubscribeUpdateAccountInfo) Reset() {
 	*x = SubscribeUpdateAccountInfo{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[13]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[13]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateAccountInfo) String() string {
@@ -1119,7 +1091,7 @@ func (*SubscribeUpdateAccountInfo) ProtoMessage() {}
 
 func (x *SubscribeUpdateAccountInfo) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[13]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1202,11 +1174,9 @@ type SubscribeUpdateSlot struct {
 
 func (x *SubscribeUpdateSlot) Reset() {
 	*x = SubscribeUpdateSlot{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[14]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[14]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateSlot) String() string {
@@ -1217,7 +1187,7 @@ func (*SubscribeUpdateSlot) ProtoMessage() {}
 
 func (x *SubscribeUpdateSlot) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[14]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1264,11 +1234,9 @@ type SubscribeUpdateTransaction struct {
 
 func (x *SubscribeUpdateTransaction) Reset() {
 	*x = SubscribeUpdateTransaction{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[15]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[15]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateTransaction) String() string {
@@ -1279,7 +1247,7 @@ func (*SubscribeUpdateTransaction) ProtoMessage() {}
 
 func (x *SubscribeUpdateTransaction) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[15]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1322,11 +1290,9 @@ type SubscribeUpdateTransactionInfo struct {
 
 func (x *SubscribeUpdateTransactionInfo) Reset() {
 	*x = SubscribeUpdateTransactionInfo{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[16]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[16]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateTransactionInfo) String() string {
@@ -1337,7 +1303,7 @@ func (*SubscribeUpdateTransactionInfo) ProtoMessage() {}
 
 func (x *SubscribeUpdateTransactionInfo) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[16]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1401,11 +1367,9 @@ type SubscribeUpdateTransactionStatus struct {
 
 func (x *SubscribeUpdateTransactionStatus) Reset() {
 	*x = SubscribeUpdateTransactionStatus{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[17]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[17]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateTransactionStatus) String() string {
@@ -1416,7 +1380,7 @@ func (*SubscribeUpdateTransactionStatus) ProtoMessage() {}
 
 func (x *SubscribeUpdateTransactionStatus) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[17]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1488,11 +1452,9 @@ type SubscribeUpdateBlock struct {
 
 func (x *SubscribeUpdateBlock) Reset() {
 	*x = SubscribeUpdateBlock{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[18]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[18]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateBlock) String() string {
@@ -1503,7 +1465,7 @@ func (*SubscribeUpdateBlock) ProtoMessage() {}
 
 func (x *SubscribeUpdateBlock) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[18]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1627,11 +1589,9 @@ type SubscribeUpdateBlockMeta struct {
 
 func (x *SubscribeUpdateBlockMeta) Reset() {
 	*x = SubscribeUpdateBlockMeta{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[19]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[19]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateBlockMeta) String() string {
@@ -1642,7 +1602,7 @@ func (*SubscribeUpdateBlockMeta) ProtoMessage() {}
 
 func (x *SubscribeUpdateBlockMeta) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[19]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1735,11 +1695,9 @@ type SubscribeUpdateEntry struct {
 
 func (x *SubscribeUpdateEntry) Reset() {
 	*x = SubscribeUpdateEntry{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[20]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[20]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdateEntry) String() string {
@@ -1750,7 +1708,7 @@ func (*SubscribeUpdateEntry) ProtoMessage() {}
 
 func (x *SubscribeUpdateEntry) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[20]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1815,11 +1773,9 @@ type SubscribeUpdatePing struct {
 
 func (x *SubscribeUpdatePing) Reset() {
 	*x = SubscribeUpdatePing{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[21]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[21]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdatePing) String() string {
@@ -1830,7 +1786,7 @@ func (*SubscribeUpdatePing) ProtoMessage() {}
 
 func (x *SubscribeUpdatePing) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[21]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1855,11 +1811,9 @@ type SubscribeUpdatePong struct {
 
 func (x *SubscribeUpdatePong) Reset() {
 	*x = SubscribeUpdatePong{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[22]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[22]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SubscribeUpdatePong) String() string {
@@ -1870,7 +1824,7 @@ func (*SubscribeUpdatePong) ProtoMessage() {}
 
 func (x *SubscribeUpdatePong) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[22]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1902,11 +1856,9 @@ type PingRequest struct {
 
 func (x *PingRequest) Reset() {
 	*x = PingRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[23]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[23]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PingRequest) String() string {
@@ -1917,7 +1869,7 @@ func (*PingRequest) ProtoMessage() {}
 
 func (x *PingRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[23]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1949,11 +1901,9 @@ type PongResponse struct {
 
 func (x *PongResponse) Reset() {
 	*x = PongResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[24]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[24]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PongResponse) String() string {
@@ -1964,7 +1914,7 @@ func (*PongResponse) ProtoMessage() {}
 
 func (x *PongResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[24]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1996,11 +1946,9 @@ type GetLatestBlockhashRequest struct {
 
 func (x *GetLatestBlockhashRequest) Reset() {
 	*x = GetLatestBlockhashRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[25]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[25]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetLatestBlockhashRequest) String() string {
@@ -2011,7 +1959,7 @@ func (*GetLatestBlockhashRequest) ProtoMessage() {}
 
 func (x *GetLatestBlockhashRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[25]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2045,11 +1993,9 @@ type GetLatestBlockhashResponse struct {
 
 func (x *GetLatestBlockhashResponse) Reset() {
 	*x = GetLatestBlockhashResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[26]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[26]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetLatestBlockhashResponse) String() string {
@@ -2060,7 +2006,7 @@ func (*GetLatestBlockhashResponse) ProtoMessage() {}
 
 func (x *GetLatestBlockhashResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[26]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2106,11 +2052,9 @@ type GetBlockHeightRequest struct {
 
 func (x *GetBlockHeightRequest) Reset() {
 	*x = GetBlockHeightRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[27]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[27]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetBlockHeightRequest) String() string {
@@ -2121,7 +2065,7 @@ func (*GetBlockHeightRequest) ProtoMessage() {}
 
 func (x *GetBlockHeightRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[27]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2153,11 +2097,9 @@ type GetBlockHeightResponse struct {
 
 func (x *GetBlockHeightResponse) Reset() {
 	*x = GetBlockHeightResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[28]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[28]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetBlockHeightResponse) String() string {
@@ -2168,7 +2110,7 @@ func (*GetBlockHeightResponse) ProtoMessage() {}
 
 func (x *GetBlockHeightResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[28]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2200,11 +2142,9 @@ type GetSlotRequest struct {
 
 func (x *GetSlotRequest) Reset() {
 	*x = GetSlotRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[29]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[29]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetSlotRequest) String() string {
@@ -2215,7 +2155,7 @@ func (*GetSlotRequest) ProtoMessage() {}
 
 func (x *GetSlotRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[29]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2247,11 +2187,9 @@ type GetSlotResponse struct {
 
 func (x *GetSlotResponse) Reset() {
 	*x = GetSlotResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[30]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[30]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetSlotResponse) String() string {
@@ -2262,7 +2200,7 @@ func (*GetSlotResponse) ProtoMessage() {}
 
 func (x *GetSlotResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[30]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2292,11 +2230,9 @@ type GetVersionRequest struct {
 
 func (x *GetVersionRequest) Reset() {
 	*x = GetVersionRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[31]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[31]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetVersionRequest) String() string {
@@ -2307,7 +2243,7 @@ func (*GetVersionRequest) ProtoMessage() {}
 
 func (x *GetVersionRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[31]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2332,11 +2268,9 @@ type GetVersionResponse struct {
 
 func (x *GetVersionResponse) Reset() {
 	*x = GetVersionResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[32]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[32]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetVersionResponse) String() string {
@@ -2347,7 +2281,7 @@ func (*GetVersionResponse) ProtoMessage() {}
 
 func (x *GetVersionResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[32]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2380,11 +2314,9 @@ type IsBlockhashValidRequest struct {
 
 func (x *IsBlockhashValidRequest) Reset() {
 	*x = IsBlockhashValidRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[33]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[33]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *IsBlockhashValidRequest) String() string {
@@ -2395,7 +2327,7 @@ func (*IsBlockhashValidRequest) ProtoMessage() {}
 
 func (x *IsBlockhashValidRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[33]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2435,11 +2367,9 @@ type IsBlockhashValidResponse struct {
 
 func (x *IsBlockhashValidResponse) Reset() {
 	*x = IsBlockhashValidResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_geyser_proto_msgTypes[34]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_geyser_proto_msgTypes[34]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *IsBlockhashValidResponse) String() string {
@@ -2450,7 +2380,7 @@ func (*IsBlockhashValidResponse) ProtoMessage() {}
 
 func (x *IsBlockhashValidResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_geyser_proto_msgTypes[34]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2982,7 +2912,7 @@ func file_geyser_proto_rawDescGZIP() []byte {
 
 var file_geyser_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_geyser_proto_msgTypes = make([]protoimpl.MessageInfo, 42)
-var file_geyser_proto_goTypes = []interface{}{
+var file_geyser_proto_goTypes = []any{
 	(CommitmentLevel)(0),                               // 0: geyser.CommitmentLevel
 	(*SubscribeRequest)(nil),                           // 1: geyser.SubscribeRequest
 	(*SubscribeRequestFilterAccounts)(nil),             // 2: geyser.SubscribeRequestFilterAccounts
@@ -3108,443 +3038,21 @@ func file_geyser_proto_init() {
 		return
 	}
 	file_solana_storage_proto_init()
-	if !protoimpl.UnsafeEnabled {
-		file_geyser_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestFilterAccounts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestFilterAccountsFilter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestFilterAccountsFilterMemcmp); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestFilterSlots); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestFilterTransactions); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestFilterBlocks); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestFilterBlocksMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestFilterEntry); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestAccountsDataSlice); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeRequestPing); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdate); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateAccount); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateAccountInfo); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateSlot); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateTransaction); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateTransactionInfo); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateTransactionStatus); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[18].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateBlock); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[19].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateBlockMeta); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[20].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdateEntry); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[21].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdatePing); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[22].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SubscribeUpdatePong); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[23].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PingRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[24].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PongResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[25].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetLatestBlockhashRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[26].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetLatestBlockhashResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[27].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetBlockHeightRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[28].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetBlockHeightResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[29].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetSlotRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[30].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetSlotResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[31].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetVersionRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[32].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetVersionResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[33].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*IsBlockhashValidRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_geyser_proto_msgTypes[34].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*IsBlockhashValidResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_geyser_proto_msgTypes[0].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[2].OneofWrappers = []interface{}{
+	file_geyser_proto_msgTypes[0].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[2].OneofWrappers = []any{
 		(*SubscribeRequestFilterAccountsFilter_Memcmp)(nil),
 		(*SubscribeRequestFilterAccountsFilter_Datasize)(nil),
 		(*SubscribeRequestFilterAccountsFilter_TokenAccountState)(nil),
 	}
-	file_geyser_proto_msgTypes[3].OneofWrappers = []interface{}{
+	file_geyser_proto_msgTypes[3].OneofWrappers = []any{
 		(*SubscribeRequestFilterAccountsFilterMemcmp_Bytes)(nil),
 		(*SubscribeRequestFilterAccountsFilterMemcmp_Base58)(nil),
 		(*SubscribeRequestFilterAccountsFilterMemcmp_Base64)(nil),
 	}
-	file_geyser_proto_msgTypes[4].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[5].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[6].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[11].OneofWrappers = []interface{}{
+	file_geyser_proto_msgTypes[4].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[5].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[6].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[11].OneofWrappers = []any{
 		(*SubscribeUpdate_Account)(nil),
 		(*SubscribeUpdate_Slot)(nil),
 		(*SubscribeUpdate_Transaction)(nil),
@@ -3555,12 +3063,12 @@ func file_geyser_proto_init() {
 		(*SubscribeUpdate_BlockMeta)(nil),
 		(*SubscribeUpdate_Entry)(nil),
 	}
-	file_geyser_proto_msgTypes[13].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[14].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[25].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[27].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[29].OneofWrappers = []interface{}{}
-	file_geyser_proto_msgTypes[33].OneofWrappers = []interface{}{}
+	file_geyser_proto_msgTypes[13].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[14].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[25].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[27].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[29].OneofWrappers = []any{}
+	file_geyser_proto_msgTypes[33].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
