@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.5.1
-// - protoc             v3.19.6
+// - protoc             v5.28.3
 // source: geyser.proto
 
 package proto
